refactor: ExtractClothes 위치 변경
